LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/select-option-hover-styles-expected.txt

one two disabled

FAIL Hover styles should be present for appearance:base-select options. assert_equals: Option background-color while hovering. expected "lab(0 0 0 / 0.1)" but got "rgba(0, 0, 0, 0)"
